requirements.txt
fastapi==0.128.0
pydantic==2.12.5
pydantic-settings==2.5.0
httpx==0.28.1
uvicorn==0.40.0
uvicorn[standard]
Jinja2==3.1.6
langchain==1.2.1
langchain-google-genai==4.1.3
streamlit==1.34.0
plotly==6.5.0
langdetect==1.0.9
mkdocs==1.6.1
mkdocs-material==9.6.18
pytest==7.4.4
python-dotenv==1.0.0
nltk==3.9.1
scikit-learn==1.5.0
numpy==1.26.4
matplotlib==3.9.0
requests==2.32.5
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.21
